backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.3.0
SQLAlchemy>=2.0.35
python-dotenv==1.0.0
Pillow>=10.2.0
bcrypt>=4.0.0
PyPDF2==3.0.1
aiofiles==23.2.1
